Split `Query` into `Reader` and `Writer`. Closes #576.
