+++ tiledb/sm/kv/kv.cc
@@ -624,10 +624,9 @@ Status KV::realloc_read_buffers() {
 
 Status KV::submit_read_query(const uint64_t* subarray) {
   // Create and send query
-  auto query = new Query();
-  RETURN_NOT_OK_ELSE(
-      storage_manager_->query_init(query, kv_uri_.c_str(), QueryType::READ),
-      delete query);
+  auto query = (Query*)nullptr;
+  RETURN_NOT_OK(
+      storage_manager_->query_init(&query, kv_uri_.c_str(), QueryType::READ));
   RETURN_NOT_OK_ELSE(
       query->set_buffers(
           (const char**)read_attributes_,
@@ -646,10 +645,9 @@ Status KV::submit_read_query(const uint64_t* subarray) {
 }
 
 Status KV::submit_write_query() {
-  auto query = new Query();
-  RETURN_NOT_OK_ELSE(
-      storage_manager_->query_init(query, kv_uri_.c_str(), QueryType::WRITE),
-      delete query);
+  auto query = (Query*)nullptr;
+  RETURN_NOT_OK(
+      storage_manager_->query_init(&query, kv_uri_.c_str(), QueryType::WRITE));
   RETURN_NOT_OK_ELSE(
       query->set_buffers(
           (const char**)write_attributes_,
