+++ tiledb/sm/misc/comparators.h
@@ -38,7 +38,7 @@
 #include <cinttypes>
 #include <vector>
 
-#include "tiledb/sm/query/query.h"
+#include "tiledb/sm/query/reader.h"
 
 namespace tiledb {
 namespace sm {
@@ -64,8 +64,8 @@ class RowCmp {
    * @return `true` if `a` precedes `b` and `false` otherwise.
    */
   bool operator()(
-      const std::shared_ptr<Query::OverlappingCoords<T>>& a,
-      const std::shared_ptr<Query::OverlappingCoords<T>>& b) {
+      const std::shared_ptr<Reader::OverlappingCoords<T>>& a,
+      const std::shared_ptr<Reader::OverlappingCoords<T>>& b) {
     for (unsigned int i = 0; i < dim_num_; ++i) {
       if (a->coords_[i] < b->coords_[i])
         return true;
@@ -103,8 +103,8 @@ class ColCmp {
    * @return `true` if `a` precedes `b` and `false` otherwise.
    */
   bool operator()(
-      const std::shared_ptr<Query::OverlappingCoords<T>>& a,
-      const std::shared_ptr<Query::OverlappingCoords<T>>& b) {
+      const std::shared_ptr<Reader::OverlappingCoords<T>>& a,
+      const std::shared_ptr<Reader::OverlappingCoords<T>>& b) {
     for (unsigned int i = dim_num_ - 1;; --i) {
       if (a->coords_[i] < b->coords_[i])
         return true;
@@ -150,8 +150,8 @@ class GlobalCmp {
    * @return `true` if `a` precedes `b` and `false` otherwise.
    */
   bool operator()(
-      const std::shared_ptr<Query::OverlappingCoords<T>>& a,
-      const std::shared_ptr<Query::OverlappingCoords<T>>& b) {
+      const std::shared_ptr<Reader::OverlappingCoords<T>>& a,
+      const std::shared_ptr<Reader::OverlappingCoords<T>>& b) {
     // Compare tile order first
     auto tile_cmp = domain_->tile_order_cmp<T>(a->coords_, b->coords_);
 
@@ -215,7 +215,7 @@ class DenseCellRangeCmp {
    * @return `true` if `a` succeeds `b` and `false` otherwise.
    */
   bool operator()(
-      const Query::DenseCellRange<T>& a, const Query::DenseCellRange<T>& b) {
+      const Reader::DenseCellRange<T>& a, const Reader::DenseCellRange<T>& b) {
     if (a.start_ < b.start_)
       return false;
     if (a.start_ > b.start_)
