+++ tiledb/sm/query/query.h
@@ -38,208 +38,25 @@
 #include "tiledb/sm/fragment/fragment_metadata.h"
 #include "tiledb/sm/misc/status.h"
 #include "tiledb/sm/query/dense_cell_range_iter.h"
+#include "tiledb/sm/query/reader.h"
+#include "tiledb/sm/query/writer.h"
 #include "tiledb/sm/storage_manager/storage_manager.h"
-#include "tiledb/sm/tile/tile.h"
 
 #include <functional>
-#include <memory>
-#include <unordered_map>
 #include <vector>
 
 namespace tiledb {
 namespace sm {
 
-class StorageManager;
-
 /** Processes a (read/write) query. */
 class Query {
  public:
-  /* ********************************* */
-  /*          TYPE DEFINITIONS         */
-  /* ********************************* */
-
-  struct AttributeBuffer {
-    /**
-     * The attribute buffer. In case the attribute is var-sized, this is
-     * the offsets buffer.
-     */
-    void* buffer_;
-    /**
-     * For a var-sized attribute, this is the data buffer. It is `nullptr`
-     * for fixed-sized attributes.
-     */
-    void* buffer_var_;
-    /** The size (in bytes) of `buffer_`. */
-    uint64_t* buffer_size_;
-    /** The size (in bytes) of `buffer_var_`. */
-    uint64_t* buffer_var_size_;
-
-    /** Constructor. */
-    AttributeBuffer(
-        void* buffer,
-        void* buffer_var,
-        uint64_t* buffer_size,
-        uint64_t* buffer_var_size)
-        : buffer_(buffer)
-        , buffer_var_(buffer_var)
-        , buffer_size_(buffer_size)
-        , buffer_var_size_(buffer_var_size) {
-    }
-  };
-
-  struct GlobalWriteState {
-    /**
-     * Stores the last tile of each attribute for each write operation.
-     * For fixed-sized attributes, the second tile is ignored. For
-     * var-sized attributes, the first tile is the offsets tile, whereas
-     * the second tile is the values tile.
-     */
-    std::unordered_map<std::string, std::pair<Tile, Tile>> last_tiles_;
-
-    /**
-     * Stores the number of cells written for each attribute across the
-     * write operations.
-     */
-    std::unordered_map<std::string, uint64_t> cells_written_;
-
-    /** The fragment metadata. */
-    std::shared_ptr<FragmentMetadata> frag_meta_;
-  };
-
-  /**
-   * For each fixed-sized attributes, the second tile in the pair is
-   * ignored. For var-sized attributes, the first is a pointer to the
-   * offsets tile and the second is a pointer to the var-sized values tile.
-   */
-  typedef std::pair<std::shared_ptr<Tile>, std::shared_ptr<Tile>> TilePair;
-
-  /** Information about a tile (across multiple attributes). */
-  struct OverlappingTile {
-    /** A fragment index. */
-    unsigned fragment_idx_;
-    /** The tile index in the fragment. */
-    uint64_t tile_idx_;
-    /** `true` if the overlap is full, and `false` if it is partial. */
-    bool full_overlap_;
-    /**
-     * Maps attribute names to attribute tiles. Note that the coordinates
-     * are a special attribute as well.
-     */
-    std::unordered_map<std::string, TilePair> attr_tiles_;
-
-    /** Constructor. */
-    OverlappingTile(
-        unsigned fragment_idx, uint64_t tile_idx, bool full_overlap = false)
-        : fragment_idx_(fragment_idx)
-        , tile_idx_(tile_idx)
-        , full_overlap_(full_overlap) {
-    }
-  };
-
-  /** A vector of overlapping tiles. */
-  typedef std::vector<std::shared_ptr<OverlappingTile>> OverlappingTileVec;
-
-  /** A cell range belonging to a particular overlapping tile. */
-  struct OverlappingCellRange {
-    /**
-     * The tile the cell range belongs to. If `nullptr`, then this is
-     * an "empty" cell range, to be filled with the default empty
-     * values.
-     */
-    std::shared_ptr<OverlappingTile> tile_;
-    /** The starting cell in the range. */
-    uint64_t start_;
-    /** The ending cell in the range. */
-    uint64_t end_;
-
-    /** Constructor. */
-    OverlappingCellRange(
-        std::shared_ptr<OverlappingTile> tile, uint64_t start, uint64_t end)
-        : tile_(std::move(tile))
-        , start_(start)
-        , end_(end) {
-    }
-  };
-
-  /** A list of cell ranges. */
-  typedef std::list<std::shared_ptr<OverlappingCellRange>>
-      OverlappingCellRangeList;
-
-  /**
-   * Records the overlapping tile and position of the coordinates
-   * in that tile.
-   *
-   * @tparam T The coords type
-   */
-  template <class T>
-  struct OverlappingCoords {
-    /** The overlapping tile the coords belong to. */
-    std::shared_ptr<OverlappingTile> tile_;
-    /** The coordinates. */
-    const T* coords_;
-    /** The position of the coordinates in the tile. */
-    uint64_t pos_;
-
-    /** Constructor. */
-    OverlappingCoords(
-        std::shared_ptr<OverlappingTile> tile, const T* coords, uint64_t pos)
-        : tile_(std::move(tile))
-        , coords_(coords)
-        , pos_(pos) {
-    }
-  };
-
-  /** A cell range produced by the dense read algorithm. */
-  template <class T>
-  struct DenseCellRange {
-    /**
-     * The fragment index. `-1` stands for no fragment, which means
-     * that the cell range must be filled with the fill value.
-     */
-    int fragment_idx_;
-    /** The tile coordinates of the range. */
-    const T* tile_coords_;
-    /** The starting cell in the range. */
-    uint64_t start_;
-    /** The ending cell in the range. */
-    uint64_t end_;
-
-    /** Constructor. */
-    DenseCellRange(
-        int fragment_idx, const T* tile_coords, uint64_t start, uint64_t end)
-        : fragment_idx_(fragment_idx)
-        , tile_coords_(tile_coords)
-        , start_(start)
-        , end_(end) {
-    }
-  };
-
-  /** Cell range to be written. */
-  struct WriteCellRange {
-    /** The position in the tile where the range will be copied. */
-    uint64_t pos_;
-    /** The starting cell in the user buffers. */
-    uint64_t start_;
-    /** The ending cell in the user buffers. */
-    uint64_t end_;
-
-    /** Constructor. */
-    WriteCellRange(uint64_t pos, uint64_t start, uint64_t end)
-        : pos_(pos)
-        , start_(start)
-        , end_(end) {
-    }
-  };
-
-  /** A vector of write cell ranges. */
-  typedef std::vector<WriteCellRange> WriteCellRangeVec;
-
   /* ********************************* */
   /*     CONSTRUCTORS & DESTRUCTORS    */
   /* ********************************* */
 
   /** Constructor. */
-  Query();
+  explicit Query(QueryType type);
 
   /** Destructor. */
   ~Query();
@@ -248,165 +65,12 @@ class Query {
   /*                 API               */
   /* ********************************* */
 
-  /**
-   * Computes info about the overlapping tiles, such as which fragment they
-   * belong to, the tile index and the type of overlap.
-   *
-   * @tparam T The coords type.
-   * @param tiles The tiles to be computed.
-   * @return Status
-   */
-  template <class T>
-  Status compute_overlapping_tiles(OverlappingTileVec* tiles) const;
-
-  /**
-   * Retrieves the tiles on a particular attribute from all input fragments
-   * based on the tile info in `tiles`.
-   *
-   * @param attribute The attribute name.
-   * @param tiles The retrieved tiles will be stored in `tiles`.
-   * @return Status
-   */
-  Status read_tiles(
-      const std::string& attribute, OverlappingTileVec* tiles) const;
-
-  /**
-   * Computes the overlapping coordinates for a given subarray.
-   *
-   * @tparam T The coords type.
-   * @param tiles The tiles to get the overlapping coordinates from.
-   * @param coords The coordinates to be retrieved.
-   * @return Status
-   */
-  template <class T>
-  Status compute_overlapping_coords(
-      const OverlappingTileVec& tiles,
-      std::list<std::shared_ptr<OverlappingCoords<T>>>* coords) const;
-
-  /**
-   * Retrieves the coordinates that overlap the subarray from the input
-   * overlapping tile.
-   *
-   * @tparam T The coords type.
-   * @param The overlapping tile.
-   * @param coords The overlapping coordinates to retrieve.
-   * @return Status
-   */
-  template <class T>
-  Status compute_overlapping_coords(
-      const std::shared_ptr<OverlappingTile>& tile,
-      std::list<std::shared_ptr<OverlappingCoords<T>>>* coords) const;
-
-  /**
-   * Gets all the coordinates of the input tile into `coords`.
-   *
-   * @tparam T The coords type.
-   * @param tile The overlapping tile to read the coordinates from.
-   * @param coords The overlapping coordinates to copy into.
-   * @return Status
-   */
-  template <class T>
-  Status get_all_coords(
-      const std::shared_ptr<OverlappingTile>& tile,
-      std::list<std::shared_ptr<OverlappingCoords<T>>>* coords) const;
-
-  /**
-   * Sorts the input coordinates according to the input layout.
-   *
-   * @tparam T The coords type.
-   * @param coords The coordinates to sort.
-   * @return Status
-   */
-  template <class T>
-  Status sort_coords(
-      std::list<std::shared_ptr<OverlappingCoords<T>>>* coords) const;
-
-  /**
-   * Deduplicates the input coordinates, breaking ties giving preference
-   * to the largest fragment index (i.e., it prefers more recent fragments).
-   *
-   * @tparam T The coords type.
-   * @param coords The coordinates to dedup.
-   * @return Status
-   */
-  template <class T>
-  Status dedup_coords(
-      std::list<std::shared_ptr<OverlappingCoords<T>>>* coords) const;
-
-  /**
-   * Compute the maximal cell ranges of contiguous cell positions.
-   *
-   * @tparam T The coords type.
-   * @param coords The coordinates to compute the ranges from.
-   * @param cell_ranges The cell ranges to compute.
-   * @return Status
-   */
-  template <class T>
-  Status compute_cell_ranges(
-      const std::list<std::shared_ptr<OverlappingCoords<T>>>& coords,
-      OverlappingCellRangeList* cell_ranges) const;
-
-  /**
-   * Copies the cells for the input attribute and cell ranges, into
-   * the corresponding result buffers.
-   *
-   * @param attribute The targeted attribute.
-   * @param cell_ranges The cell ranges to copy cells for.
-   * @return Status
-   */
-  Status copy_cells(
-      const std::string& attribute,
-      const OverlappingCellRangeList& cell_ranges) const;
-
-  /**
-   * Copies the cells for the input **fixed-sized** attribute and cell
-   * ranges, into the corresponding result buffers.
-   *
-   * @param attribute The targeted attribute.
-   * @param cell_ranges The cell ranges to copy cells for.
-   * @return Status
-   */
-  Status copy_fixed_cells(
-      const std::string& attribute,
-      const OverlappingCellRangeList& cell_ranges) const;
-
-  /**
-   * Copies the cells for the input **var-sized** attribute and cell
-   * ranges, into the corresponding result buffers.
-   *
-   * @param attribute The targeted attribute.
-   * @param cell_ranges The cell ranges to copy cells for.
-   * @return Status
-   */
-  Status copy_var_cells(
-      const std::string& attribute,
-      const OverlappingCellRangeList& cell_ranges) const;
-
-  /**
-   * Checks whether two hyper-rectangles overlap, and determines whether
-   * the first rectangle contains the second.
-   *
-   * @tparam T The domain type.
-   * @param a The first rectangle.
-   * @param b The second rectangle.
-   * @param dim_num The number of dimensions.
-   * @param a_contains_b Determines whether the first rectangle contains the
-   *     second.
-   * @return `True` if the rectangles overlap, and `false` otherwise.
-   */
-  template <class T>
-  bool overlap(
-      const T* a, const T* b, unsigned dim_num, bool* a_contains_b) const;
-
-  /** Returns the array schema.*/
+  /** Returns the array schema. */
   const ArraySchema* array_schema() const;
-
-  /** Processes a query. */
-  Status process();
 
   /**
    * Computes a vector of `subarrays` into which `subarray` must be partitioned,
-   * such that each subarray in `subarrays` can be saferly answered by the
+   * such that each subarray in `subarrays` can be safely answered by the
    * query without a memory overflow.
    *
    * @param subarray The input subarray.
@@ -421,23 +85,23 @@ class Query {
    */
   Status finalize();
 
+  /** Returns the number of fragments involved in the (read) query. */
+  unsigned fragment_num() const;
+
   /** Returns a vector with the fragment URIs. */
   std::vector<URI> fragment_uris() const;
 
-  /** Returns the number of fragments involved in the query. */
-  unsigned int fragment_num() const;
-
-  /**
-   * Initializes the query states. This must be called before the query is
-   * submitted.
-   */
+  /** Initializes the query. */
   Status init();
 
-  /** Returns the lastly created fragment uri. */
+  /** Returns the last fragment uri. */
   URI last_fragment_uri() const;
 
   /** Returns the cell layout. */
   Layout layout() const;
+
+  /** Processes a query. */
+  Status process();
 
   /** Sets the array schema. */
   void set_array_schema(const ArraySchema* array_schema);
@@ -475,14 +139,14 @@ class Query {
   void set_callback(
       const std::function<void(void*)>& callback, void* callback_data);
 
-  /** Sets and initializes the fragment metadata. */
-  Status set_fragment_metadata(
+  /** Sets the fragment metadata. */
+  void set_fragment_metadata(
       const std::vector<FragmentMetadata*>& fragment_metadata);
 
   /**
    * Sets the cell layout of the query. The function will return an error
    * if the queried array is a key-value store (because it has its default
-   * layout.
+   * layout for both reads and writes.
    */
   Status set_layout(Layout layout);
 
@@ -498,17 +162,8 @@ class Query {
    */
   Status set_subarray(const void* subarray);
 
-  /** Sets the query type. */
-  void set_type(QueryType type);
-
   /** Returns the query status. */
   QueryStatus status() const;
-
-  /** Returns the storage manager. */
-  StorageManager* storage_manager() const;
-
-  /** Returns the subarray in which the query is constrained. */
-  const void* subarray() const;
 
   /** Returns the query type. */
   QueryType type() const;
@@ -517,18 +172,6 @@ class Query {
   /* ********************************* */
   /*         PRIVATE ATTRIBUTES        */
   /* ********************************* */
-
-  /** The state associated with global writes. */
-  std::unique_ptr<GlobalWriteState> global_write_state_;
-
-  /** The names of the attributes involved in the query. */
-  std::vector<std::string> attributes_;
-
-  /** The array schema. */
-  const ArraySchema* array_schema_;
-
-  /** Maps attribute names to their buffers. */
-  std::unordered_map<std::string, AttributeBuffer> attr_buffers_;
 
   /** A function that will be called upon the completion of an async query. */
   std::function<void(void*)> callback_;
@@ -539,545 +182,18 @@ class Query {
   /** The query status. */
   QueryStatus status_;
 
-  /** The metadata of the fragments involved in the query. */
-  std::vector<FragmentMetadata*> fragment_metadata_;
-
-  /** The cell layout. */
-  Layout layout_;
-
-  /** The storage manager. */
-  StorageManager* storage_manager_;
-
-  /**
-   * The subarray the query is constrained on. A nullptr implies the
-   * entire domain.
-   */
-  void* subarray_;
-
   /** The query type. */
   QueryType type_;
+
+  /** Query reader. */
+  Reader reader_;
+
+  /** Wuery writer. */
+  Writer writer_;
 
   /* ********************************* */
   /*           PRIVATE METHODS         */
   /* ********************************* */
-
-  /** Checks if attributes has been appropriately set for a query. */
-  Status check_attributes();
-
-  /**
-   * Checks if the buffer sizes are correct in the case of writing
-   * in a dense array in an ordered layout.
-   *
-   * @return Status
-   */
-  Status check_buffer_sizes_ordered() const;
-
-  /** Checks if `subarray` falls inside the array domain. */
-  Status check_subarray(const void* subarray) const;
-
-  /** Checks if `subarray` falls inside the array domain. */
-  template <class T>
-  Status check_subarray(const T* subarray) const;
-
-  /**
-   * For the given cell range, it computes all the result dense cell ranges
-   * across fragments, given precedence to more recent fragments.
-   *
-   * @tparam T The domain type.
-   * @param tile_coords The tile coordinates in the array domain.
-   * @param frag_its The fragment dence cell range iterators.
-   * @param start The start position of the range this function focuses on.
-   * @param end The end position of the range this function focuses on.
-   * @param dense_cell_ranges The cell ranges where the results are appended to.
-   * @return Status
-   *
-   * @note The input dense cell range iterators will be appropriately
-   *     incremented.
-   */
-  template <class T>
-  Status compute_dense_cell_ranges(
-      const T* tile_coords,
-      std::vector<DenseCellRangeIter<T>>& frag_its,
-      uint64_t start,
-      uint64_t end,
-      std::list<DenseCellRange<T>>* dense_cell_ranges);
-
-  /**
-   * Computes the dense overlapping tiles and cell ranges based on the
-   * input dense cell ranges. Note that the function also computes
-   * the maximal ranges of contiguous cells for each fragment/tile pair.
-   *
-   * @tparam T The domain type.
-   * @param dense_cell_ranges The dense cell ranges the overlapping tiles
-   *     and cell ranges will be derived from.
-   * @param tiles The overlapping tiles to be computed.
-   * @param overlapping_cell_ranges The overlapping cell ranges to be
-   *     computed.
-   * @return Status
-   */
-  template <class T>
-  Status compute_dense_overlapping_tiles_and_cell_ranges(
-      const std::list<DenseCellRange<T>>& dense_cell_ranges,
-      const std::list<std::shared_ptr<OverlappingCoords<T>>>& coords,
-      OverlappingTileVec* tiles,
-      OverlappingCellRangeList* overlapping_cell_ranges);
-
-  /** Returns the empty fill value based on the input datatype. */
-  const void* fill_value(Datatype type) const;
-
-  /**
-   * Initializes the fragment dense cell range iterators. There is one vector
-   * per tile overlapping with the query subarray, which stores one cell range
-   * iterator per fragment.
-   *
-   * @tparam T The domain type.
-   * @param iters The iterators to be initialized.
-   * @param overlapping_tile_idx_coords A map from global tile index to a pair
-   *     (overlapping tile index, overlapping tile coords).
-   */
-  template <class T>
-  Status init_tile_fragment_dense_cell_range_iters(
-      std::vector<std::vector<DenseCellRangeIter<T>>>* iters,
-      std::unordered_map<uint64_t, std::pair<uint64_t, std::vector<T>>>*
-          overlapping_tile_idx_coords);
-
-  /**
-   * Returns a new fragment name, which is in the form: <br>
-   * .__thread-id_timestamp. For instance,
-   *  __6426153_1458759561320
-   *
-   * Note that this is a temporary name, initiated by a new write process.
-   * After the new fragmemt is finalized, the array will change its name
-   * by removing the leading '.' character.
-   *
-   * @return A new special fragment name on success, or "" (empty string) on
-   *     error.
-   */
-  std::string new_fragment_name() const;
-
-  /** Sets the query attributes. */
-  Status set_attributes(const char** attributes, unsigned int attribute_num);
-
-  /** Sets the buffer sizes to zero. */
-  void zero_out_buffer_sizes();
-
-  /**
-   * Initializes dense cell range iterators for the subarray to be writte,
-   * one per overlapping tile.
-   *
-   * @tparam T The domain type.
-   * @param iters The dense cell range iterators to be created.
-   * @return Status
-   */
-  template <class T>
-  Status init_tile_dense_cell_range_iters(
-      std::vector<DenseCellRangeIter<T>>* iters) const;
-
-  /**
-   * Computes the cell ranges to be written, derived from a
-   * dense cell range iterator for a specific tile.
-   *
-   * @tparam T The domain type.
-   * @param iter The dense cell range iterator for one
-   *     tile overlapping with the write subarray.
-   * @param write_cell_ranges The write cell ranges to be created.
-   * @return Status
-   */
-  template <class T>
-  Status compute_write_cell_ranges(
-      DenseCellRangeIter<T>* iters, WriteCellRangeVec* write_cell_ranges) const;
-
-  /**
-   * Sorts the coordinates of the user buffers, creating a vector with
-   * the sorted positions.
-   *
-   * @tparam T The domain type.
-   * @param cell_pos The sorted cell positions to be created.
-   * @return Status
-   */
-  template <class T>
-  Status sort_coords(std::vector<uint64_t>* cell_pos) const;
-
-  /** Initializes the global write state. */
-  Status init_global_write_state();
-
-  /** Finalizes the global write state. */
-  Status finalize_global_write_state();
-
-  /**
-   * Finalizes the global write state.
-   *
-   * @tparam T The domain type.
-   * @return Status
-   */
-  template <class T>
-  Status finalize_global_write_state();
-
-  /**
-   * Applicable only to write in global order. It prepares only full
-   * tiles, storing the last potentially non-full tile in
-   * `last_tiles_` as part of the state to be used in the next
-   * write invocation. The last tiles are written to storage
-   * upon `finalize`. Upon each invocation, the function first
-   * populates the partially full last tile from the previous
-   * invocation.
-   *
-   * @param attribute The attribute to prepare the tiles for.
-   * @param tiles The **full** tiles to be created.
-   * @return Status
-   */
-  Status prepare_full_tiles(
-      const std::string& attribute, std::vector<Tile>* tiles) const;
-
-  /**
-   * Applicable only to write in global order. It prepares only full
-   * tiles, storing the last potentially non-full tile in
-   * `last_tiles_` as part of the state to be used in the next
-   * write invocation. The last tiles are written to storage
-   * upon `finalize`. Upon each invocation, the function first
-   * populates the partially full last tile from the previous
-   * invocation. Applicable only to fixed-sized attributes.
-   *
-   * @param attribute The attribute to prepare the tiles for.
-   * @param tiles The **full** tiles to be created.
-   * @return Status
-   */
-  Status prepare_full_tiles_fixed(
-      const std::string& attribute, std::vector<Tile>* tiles) const;
-
-  /**
-   * Applicable only to write in global order. It prepares only full
-   * tiles, storing the last potentially non-full tile in
-   * `last_tiles_` as part of the state to be used in the next
-   * write invocation. The last tiles are written to storage
-   * upon `finalize`. Upon each invocation, the function first
-   * populates the partially full last tile from the previous
-   * invocation. Applicable only to var-sized attributes.
-   *
-   * @param attribute The attribute to prepare the tiles for.
-   * @param tiles The **full** tiles to be created.
-   * @return Status
-   */
-  Status prepare_full_tiles_var(
-      const std::string& attribute, std::vector<Tile>* tiles) const;
-
-  /**
-   * It prepares the tiles, copying from the user buffers into the tiles
-   * the values based on the input write cell ranges, focusing on the
-   * input attribute.
-   *
-   * @param attribute The attribute to prepare the tiles for.
-   * @param write_cell_ranges The write cell ranges.
-   * @param tiles The tiles to be created.
-   * @return Status
-   */
-  Status prepare_tiles(
-      const std::string& attribute,
-      const std::vector<WriteCellRangeVec>& write_cell_ranges,
-      std::vector<Tile>* tiles) const;
-
-  /**
-   * It prepares the tiles, re-organizing the cells from the user
-   * buffers based on the input sorted positions.
-   *
-   * @param attribute The attribute to prepare the tiles for.
-   * @param cell_pos The positions that resulted from sorting and
-   *     according to which the cells must be re-arranged.
-   * @param tiles The tiles to be created.
-   * @return Status
-   */
-  Status prepare_tiles(
-      const std::string& attribute,
-      const std::vector<uint64_t>& cell_pos,
-      std::vector<Tile>* tiles) const;
-
-  /**
-   * It prepares the tiles, re-organizing the cells from the user
-   * buffers based on the input sorted positions. Applicable only
-   * to fixed-sized attributes.
-   *
-   * @param attribute The attribute to prepare the tiles for.
-   * @param cell_pos The positions that resulted from sorting and
-   *     according to which the cells must be re-arranged.
-   * @param tiles The tiles to be created.
-   * @return Status
-   */
-  Status prepare_tiles_fixed(
-      const std::string& attribute,
-      const std::vector<uint64_t>& cell_pos,
-      std::vector<Tile>* tiles) const;
-
-  /**
-   * It prepares the tiles, re-organizing the cells from the user
-   * buffers based on the input sorted positions. Applicable only
-   * to var-sized attributes.
-   *
-   * @param attribute The attribute to prepare the tiles for.
-   * @param cell_pos The positions that resulted from sorting and
-   *     according to which the cells must be re-arranged.
-   * @param tiles The tiles to be created.
-   * @return Status
-   */
-  Status prepare_tiles_var(
-      const std::string& attribute,
-      const std::vector<uint64_t>& cell_pos,
-      std::vector<Tile>* tiles) const;
-
-  /**
-   * Computes the coordinates metadata (e.g., MBRs).
-   *
-   * @tparam T The domain type.
-   * @param tiles The tiles to calculate the coords metadata from.
-   * @param meta The fragment metadata that will store the coords metadata.
-   * @return Status
-   */
-  template <class T>
-  Status compute_coords_metadata(
-      const std::vector<Tile>& tiles, FragmentMetadata* meta) const;
-
-  /**
-   * Writes the input tiles for the input attribute to storage.
-   *
-   * @param attribute The attribute the tiles belong to.
-   * @param frag_meta The fragment metadata.
-   * @param tiles The tiles to be written.
-   * @return Status
-   */
-  Status write_tiles(
-      const std::string& attribute,
-      FragmentMetadata* frag_meta,
-      std::vector<Tile>& tiles) const;
-
-  /**
-   * Initializes the tiles for writing for the input attribute.
-   *
-   * @param attribute The attribute the tiles belong to.
-   * @param tile_num The number of tiles.
-   * @param tiles The tiles to be initialized. Note that the vector
-   *     has been already preallocated.
-   * @return Status
-   */
-  Status init_tiles(
-      const std::string& attribute,
-      uint64_t tile_num,
-      std::vector<Tile>* tiles) const;
-
-  /**
-   * Writes the input cell range to the input tile, for a particular
-   * buffer. Applicable to **fixed-sized** attributes.
-   *
-   * @param buff The write buffer where the cells will be copied from.
-   * @param start The start element in the write buffer.
-   * @param end The end element in the write buffer.
-   * @param tile The tile to write to.
-   * @return Status
-   */
-  Status write_cell_range_to_tile(
-      ConstBuffer* buff, uint64_t start, uint64_t end, Tile* tile) const;
-
-  /**
-   * Writes the input cell range to the input tile, for a particular
-   * buffer. Applicable to **variable-sized** attributes.
-   *
-   * @param buff The write buffer where the cell offsets will be copied from.
-   * @param buff_var The write buffer where the cell values will be copied from.
-   * @param start The start element in the write buffer.
-   * @param end The end element in the write buffer.
-   * @param tile The tile offsets to write to.
-   * @param tile_var The tile with the var-sized cells to write to.
-   * @return Status
-   */
-  Status write_cell_range_to_tile_var(
-      ConstBuffer* buff,
-      ConstBuffer* buff_var,
-      uint64_t start,
-      uint64_t end,
-      Tile* tile,
-      Tile* tile_var) const;
-
-  /**
-   * Applicable only to global writes. Writes the last tiles for each
-   * attribute remaining in the state, and records the metadata for
-   * the coordinates attribute (if present).
-   *
-   * @tparam T The domain type.
-   * @return Status
-   */
-  template <class T>
-  Status global_write_handle_last_tile();
-
-  /**
-   * Writes an empty cell range to the input tile.
-   * Applicable to **fixed-sized** attributes.
-   *
-   * @param num Number of empty values to write.
-   * @param tile The tile to write to.
-   * @return Status
-   */
-  Status write_empty_cell_range_to_tile(uint64_t num, Tile* tile) const;
-
-  /**
-   * Writes an empty cell range to the input tile.
-   * Applicable to **variable-sized** attributes.
-   *
-   * @param num Number of empty values to write.
-   * @param tile The tile offsets to write to.
-   * @param tile_var The tile with the var-sized cells to write to.
-   * @return Status
-   */
-  Status write_empty_cell_range_to_tile_var(
-      uint64_t num, Tile* tile, Tile* tile_var) const;
-
-  /**
-   * Creates a new fragment.
-   *
-   * @param dense Whether the fragment is dense or not.
-   * @param frag_meta The fragment metadata to be generated.
-   * @return Status
-   */
-  Status create_fragment(
-      bool dense, std::shared_ptr<FragmentMetadata>* frag_meta) const;
-
-  /** Returns `true` if the coordinates are included in the attributes. */
-  bool has_coords() const;
-
-  /** Closes all attribute files, flushing their state to storage. */
-  Status close_files(FragmentMetadata* meta) const;
-
-  /** Perform a dense read */
-  Status dense_read();
-
-  /**
-   * Perform a dense read.
-   *
-   * @tparam The domain type.
-   * @return Status
-   */
-  template <class T>
-  Status dense_read();
-
-  /**
-   * Writes in the global layout. Applicable to both dense and sparse
-   * arrays.
-   */
-  Status global_write();
-
-  /**
-   * Writes in the global layout. Applicable to both dense and sparse
-   * arrays.
-   *
-   * @tparam T The domain type.
-   */
-  template <class T>
-  Status global_write();
-
-  /**
-   * Writes in an ordered layout (col- or row-major order). Applicable only
-   * to dense arrays.
-   */
-  Status ordered_write();
-
-  /**
-   * Writes in an ordered layout (col- or row-major order). Applicable only
-   * to dense arrays.
-   *
-   * @tparam T The domain type.
-   */
-  template <class T>
-  Status ordered_write();
-
-  /**
-   * Writes in an unordered layout. Applicable to both dense and sparse arrays.
-   * Explicit coordinates must be provided for this write.
-   */
-  Status unordered_write();
-
-  /**
-   * Writes in an unordered layout. Applicable to both dense and sparse arrays.
-   * Explicit coordinates must be provided for this write.
-   *
-   * @tparam T The domain type.
-   */
-  template <class T>
-  Status unordered_write();
-
-  /** Performs a read on a sparse array. */
-  Status sparse_read();
-
-  /**
-   * Performs a read on a sparse array.
-   *
-   * @tparam The domain type.
-   * @return Status
-   */
-  template <class T>
-  Status sparse_read();
-
-  /** Executes a read query. */
-  Status read();
-
-  /** Executes a write query. */
-  Status write();
-
-  /**
-   * Initializes a fixed-sized tile.
-   *
-   * @param attribute The attribute the tile belongs to.
-   * @param tile The tile to be initialized.
-   * @return Status
-   */
-  Status init_tile(const std::string& attribute, Tile* tile) const;
-
-  /**
-   * Initializes a var-sized tile.
-   *
-   * @param attribute The attribute the tile belongs to.
-   * @param tile The offsets tile to be initialized.
-   * @param tile_var The var-sized data tile to be initialized.
-   * @return Status
-   */
-  Status init_tile(
-      const std::string& attribute, Tile* tile, Tile* tile_var) const;
-
-  /**
-   * Handles the coordinates that fall between `start` and `end`.
-   * This function will either skip the coordinates if they belong to an
-   * older fragment than that of the current dense cell range, or include them
-   * as results and split the dense cell range.
-   *
-   * @tparam T The domain type
-   * @param cur_tile The current tile.
-   * @param cur_tile_coords The current tile coordinates.
-   * @param start The start of the dense cell range.
-   * @param end The end of the dense cell range.
-   * @param coords_size The coordintes size.
-   * @param coords The list of coordinates.
-   * @param coords_it The iterator pointing at the current coordinates.
-   * @param coords_tile The tile where the current coordinates belong to.
-   * @param coords_pos The position of the current coordinates in their tile.
-   * @param coords_fidx The fragment index of the current coordinates.
-   * @param coords_tile_coords The global tile coordinates of the tile the
-   *     current cell coordinates belong to
-   * @param overlapping_cell_ranges The result cell ranges (to be updated
-   *     by inserting a dense cell range for a coordinate result, or by
-   *     splitting the current dense cell range).
-   * @return Status
-   */
-  template <class T>
-  Status handle_coords_in_dense_cell_range(
-      const std::shared_ptr<OverlappingTile>& cur_tile,
-      const T* cur_tile_coords,
-      uint64_t* start,
-      uint64_t end,
-      uint64_t coords_size,
-      const std::list<std::shared_ptr<OverlappingCoords<T>>>& coords,
-      typename std::list<std::shared_ptr<OverlappingCoords<T>>>::const_iterator*
-          coords_it,
-      std::shared_ptr<OverlappingTile>* coords_tile,
-      uint64_t* coords_pos,
-      unsigned* coords_fidx,
-      std::vector<T>* coords_tile_coords,
-      OverlappingCellRangeList* overlapping_cell_ranges) const;
 };
 
 }  // namespace sm
